Some bugfixes

+++ src/libclient/measurement/ping/ping_linux.cpp
@@ -93,7 +93,7 @@ Ping::Ping(QObject *parent)
 
 Ping::~Ping()
 {
-    if (definition->type == ping::System)
+    if (!definition.isNull() && definition->type == ping::System)
     {
         process.kill();
         process.waitForFinished(500);
